Reimplement overlay loading with tasks
